New enemy bullet prefabs, and enemy turret now shoots at player.
Changes:
- New enemy bullet prefabs.
- Changed PlayerBox in scene to have isTrigger collider.
- Turret enemy prefab now has isTrigger collider.
- Changed FireEnemyBullet to use same logic as FirePlayerBullet and added ability to choose between true and false bullets.
- HandleDeath is now always called once in both BaseEnemy and BaseTurret.
- Adjusted random placement of enemy turret in front of player when spawned with "Y" key.

Known bugs:
- Enemy keeps rotating weirdly around z-axis, even with no collision on. Position is correct, but rotation keeps changing.
- Random enemy rotation is causing the enemy to fire in weird ways.
- FollowSpline RotateTowards is not smooth.

WIP:
- FireEnemyBullet forward axis is hardcoded as well as location. The forward axis should be based on the rotation supplied to the function.

+++ Pogplant/Scripting/Scripts/BaseEnemy.cs
@@ -189,14 +189,14 @@ namespace Scripting
                         current_interval = 0;
 
                         //FireBullet(em.GetBullet(true), i);
-                        GameUtilities.FireEnemyBullet(muzzle_transforms[i].Position, muzzle_transforms[i].Rotation);
+                        GameUtilities.FireEnemyBullet(owner.id, owner.transform.Value.Position, owner.transform.Value.Rotation, true);
                         //Debug.Log("Firing true bullet");
                     }
                     else
                     {
                         ++current_interval;
                         //FireBullet(em.GetBullet(false), i);
-                        GameUtilities.FireEnemyBullet(muzzle_transforms[i].Position, muzzle_transforms[i].Rotation);
+                        GameUtilities.FireEnemyBullet(owner.id, owner.transform.Value.Position, owner.transform.Value.Rotation);
                         //Debug.Log("Firing false bullet");
                     }
                 }
@@ -368,16 +368,18 @@ namespace Scripting
 
         private void HandleDeath()
         {
-            is_alive = false;
+            if (is_alive)
+            {
+                //gameObject.GetComponent<Destructible_Actor>().HandleDeath();
+                //FirstPersonFiringSystem.Instance.RemoveEnemyFromListOfTargets(gameObject);
 
-            //gameObject.GetComponent<Destructible_Actor>().HandleDeath();
-            //FirstPersonFiringSystem.Instance.RemoveEnemyFromListOfTargets(gameObject);
+                ////destroy script so any keyed actions will not occur
+                //Destroy(this);
 
-            ////destroy script so any keyed actions will not occur
-            //Destroy(this);
-
-            em.DeleteEnemyInstance(entityID);
-            ECS.DestroyEntity(entityID);
+                em.DeleteEnemyInstance(entityID);
+                ECS.DestroyEntity(entityID);
+                is_alive = false;
+            }
         }
 
         public override void Init(ref uint _entityID)
